Build binaries for multiple platforms
