+++ .gitignore
@@ -1,3 +1,6 @@
+# Prerequisites
+*.d
+
 # Vim swap files
 .*.swp
 
@@ -6,6 +9,11 @@
 *.ko
 *.obj
 *.elf
+
+# Linker output
+*.ilk
+*.map
+*.exp
 
 # Precompiled Headers
 *.gch
@@ -33,5 +41,22 @@
 
 # Debug files
 *.dSYM/
+*.su
+*.idb
+*.pdb
 
-su-exec
+# Kernel Module Compile Results
+*.mod*
+*.cmd
+.tmp_versions/
+modules.order
+Module.symvers
+Mkfile.old
+dkms.conf
+
+# debug information files
+*.dwo
+
+# the binary
+dist/
+working/
