--- Makefile
@@ -1,17 +1,0 @@
-
-CFLAGS ?= -Wall -Werror -g
-LDFLAGS ?=
-
-PROG := su-exec
-SRCS := $(PROG).c
-
-all: $(PROG)
-
-$(PROG): $(SRCS)
-	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS)
-
-$(PROG)-static: $(SRCS)
-	$(CC) $(CFLAGS) -o $@ $^ -static $(LDFLAGS)
-
-clean:
-	rm -f $(PROG) $(PROG)-static
