+++ test-results.txt
@@ -1,0 +1,48 @@
+Tue Aug 19 12:01:05 JST 2025
+alpine:3.9 OK
+alpine:3.10 OK
+alpine:3.11 OK
+alpine:3.12 OK
+alpine:3.13 OK
+alpine:3.14 OK
+alpine:3.15 OK
+alpine:3.16 OK
+alpine:3.17 OK
+alpine:3.18 OK
+alpine:3.19 OK
+alpine:3.20 OK
+alpine:3.21 OK
+alpine:3.22 OK
+alpine:3 OK
+alpine:latest OK
+alpine:edge OK
+debian:jessie OK
+debian:stretch OK
+debian:buster OK
+debian:bullseye OK
+debian:bookworm OK
+debian:trixie OK
+debian:forky OK
+debian:latest OK
+ubuntu:14.04 OK
+ubuntu:16.04 OK
+ubuntu:17.04 OK
+ubuntu:17.10 OK
+ubuntu:18.04 OK
+ubuntu:18.10 OK
+ubuntu:19.04 OK
+ubuntu:19.10 OK
+ubuntu:20.04 OK
+ubuntu:20.10 OK
+ubuntu:21.04 OK
+ubuntu:21.10 OK
+ubuntu:22.04 OK
+ubuntu:22.10 OK
+ubuntu:23.04 OK
+ubuntu:23.10 OK
+ubuntu:24.04 OK
+ubuntu:24.10 OK
+ubuntu:25.04 OK
+ubuntu:25.10 OK
+ubuntu:latest OK
+ubuntu:rolling OK
